% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_bma.R
\name{mr_bma}
\alias{mr_bma}
\title{Perform Mendelian Randomization Bayesian Model Averaging}
\usage{
mr_bma(
  harmonized_data,
  kmin = 5,
  kmax = 5,
  prior_prob = 0.5,
  prior_sigma = 0.5,
  top = 10,
  nrepeat = 1e+05
)
}
\arguments{
\item{harmonized_data}{(list) Harmonized data of exposures/outcomes of interest, generated using \code{\link[TwoSampleMR:mv_harmonise_data]{TwoSampleMR::mv_harmonise_data()}}}

\item{nrepeat}{(numeric) Number of permutations used when estimating Nyholt-corrected False Discovery Rate}
}
\description{
This function takes harmonized exposure/outcome data and runs Mendelian Randomization Bayesian Model Averaging to identify and prioritize relationships between the exposures and outcome.
}
\examples{
lipid_exposures <- TwoSampleMR::mv_extract_exposures(id_exposure = c("ieu-a-299", "ieu-a-300", "ieu-a-302"))
cad_outcome <- TwoSampleMR::extract_outcome_data(snps = lipid_exposures$SNP, outcomes = "ebi-a-GCST005195")
lipids_cad_harmonized <- TwoSampleMR::mv_harmonise_data(lipid_exposures, cad_outcome)
}
