% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBMA-package.R
\docType{package}
\name{MRBMA-package}
\alias{MRBMA-package}
\alias{MRBMA}
\title{MRBMA: Bayesian model averaging for multivariable Mendelian randomization}
\description{
Implements the MR-BMA method of \insertCite{Zuber2020;textual}{MRBMA}
for identifying causal exposures from genome-wide summary statistics using Bayesian model averaging. Two search strategies are
provided: exhaustive enumeration of all \eqn{2^d - 1} models via \code{MRBMA_exhaustive} (feasible for \eqn{d \leq 12} exposures)
and shotgun stochastic search via \code{MRBMA_SSS} for larger model spaces. Empirical p-values are obtained through the permutation
procedure of \insertCite{Levin2021;textual}{MRBMA}. The stochastic search algorithm is based on the shotgun stochastic search (SSS)
of \insertCite{Hans2007;textual}{MRBMA}. All inner-loop matrix computations are implemented in C++ via RcppArmadillo, with the SSS
hash table implemented as a \code{std::unordered_map} replacing the R \code{hash} package.
}
\details{
The main functions are:
\describe{
\item{\code{\link{MRBMA_exhaustive}}}{Exhaustive BMA over all \eqn{2^d - 1}
models}
\item{\code{\link{MRBMA_SSS}}}{Stochastic search BMA for large model spaces}
\item{\code{\link{report_best_models}}}{Top-K model table}
\item{\code{\link{report_MRBMA}}}{Exposure MPPI table}
\item{\code{\link{MRBMA_diagnostics}}}{Cook's distance and Q-statistic for
influential/outlying instruments}
\item{\code{\link{MRBMA_permutations}}}{Permutation null distribution}
\item{\code{\link{MRBMA_pvalues}}}{Empirical p-values with BH-FDR correction}
}
}
\references{
\insertAllCited{}
}
\author{
Maintainer: Leonardo Bottolo \href{mailto:lb664@cam.ac.uk}{lb664@cam.ac.uk}

Authors: Verena Zuber \href{mailto:v.zuber@imperial.ac.uk}{v.zuber@imperial.ac.uk}
}
