% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbma-package.R
\docType{package}
\name{mrbma-package}
\alias{mrbma}
\alias{mrbma-package}
\title{mrbma: Mendelian Randomization Bayesian Model Averaging}
\description{
A set of functions designed to implement Mendelian Randomization Bayesian Model Averaging.
}
\author{
\strong{Maintainer}: Michael Levin \email{michael.levin@pennmedicine.upenn.edu} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
